requirements.txt
gradio>=4.0.0
transformers>=4.21.0
torch>=1.12.0
sentencepiece>=0.1.97
pandas>=1.5.0
Pillow>=9.0.0
numpy>=1.21.0
flask>=2.0.0
flask-cors>=4.0.0